Convert EuiCallOut to TS
