--- src/components/call_out/index.js
@@ -1,1 +1,0 @@
-export { EuiCallOut } from './call_out';
