+++ src/components/call_out/index.ts
@@ -1,0 +1,1 @@
+export { EuiCallOut, EuiCallOutProps } from './call_out';
